+++ CHANGELOG.rst
@@ -1,6 +1,21 @@
 Change Log
 ##########
 
+Version 0.2.0 (2024-02-15)
+**********
+
+* Remove unused github actions and tests.
+
+Version 0.1.3 (2023-09-20)
+**********************************************
+
+* Fix: Fix error in logging unsuccessful responses
+
+Version 0.1.0 (2023-09-19)
+**********************************************
+
+* Feat: Add account registration and registration validation API calls
+
 Version 0.0.3 (2023-09-06)
 **********************************************